+++ src/main.py
@@ -3,25 +3,16 @@
 # Author: William Andersson <contact.kiwipy@gmail.com>
 
 import sys
-import sqlite3
+
 from PySide6 import QtGui
 from PySide6.QtWidgets import QApplication, QMainWindow, QMessageBox, QLabel
 from PySide6.QtCore import QFile, QTranslator, QLocale, QLibraryInfo
+
+import resources
+import db_view
 from window import Ui_MainWindow
 from label_view import update_label_values
-from db_view import manage_db
 from app_path import local_path_for
-import resources
-
-NEW_DB = """CREATE TABLE IF NOT EXISTS products (
-            Product TEXT PRIMARY KEY,
-            Fat NUMERIC,
-            Saturated NUMERIC,
-            Carbs NUMERIC,
-            Sugar NUMERIC,
-            Protein NUMERIC,
-            Salt NUMERIC,
-            Fiber NUMERIC);"""
 
 class MainWindow(QMainWindow):
     def __init__(self, *args, **kwargs):
@@ -55,29 +46,11 @@ class MainWindow(QMainWindow):
                 self.ComboBox_tuple[index].activated.connect(lambda checked, index=index: self.ComboBox_tuple[index+1].setEnabled(True))
 
         self.ui.clear_btn.clicked.connect(self.clear)
-        self.ui.save_btn.clicked.connect(self.save)
-        self.ui.base_btn.clicked.connect(lambda: manage_db(self))
-        
+        self.ui.save_btn.clicked.connect(lambda: db_view.db_save(self))
+        self.ui.base_btn.clicked.connect(lambda: db_view.db_manage(self))
+
         self.db_store = local_path_for(kwargs['application']) / "default.db"
-        if self.db_store != None:
-            if not self.db_store.exists():
-                try:
-                    self.database = sqlite3.connect(self.db_store)
-                    self.cursor = self.database.cursor()
-                    self.cursor.execute(NEW_DB)
-                    self.database.commit()
-                except Exception as e:
-                    print(f"Error: {e}")
-                    self.err_msg(self.tr("Could not create database."))
-            else:
-                try:
-                    self.database = sqlite3.connect(self.db_store)
-                    self.cursor = self.database.cursor()
-                except Exception as e:
-                    print(f"Error: {e}")
-                    self.err_msg(self.tr("Could not load database."))
-        else:
-            self.err_msg(self.tr("No path for database."))
+        db_view.db_load(self)
         
         for spinbox in self.SpinBox_tuple:
         # When the value off all QSpinBox changes.
@@ -126,31 +99,6 @@ class MainWindow(QMainWindow):
         self.ui.salt_value.setText(f"g")
         self.ui.fiber_value.setText(f"g")
         self.ui.energy_value.setText(f"kcal/kj")
-    
-    def save(self):
-    # Save new product to database
-        Names = self.cursor.execute(f"SELECT Product FROM products WHERE Product='{self.ui.title.text().lower()}'").fetchall()
-        if self.ui.title.text().lower() == "":
-            self.info_msg(self.tr("Missing title for product."))
-        elif len(Names) > 0 and self.ui.title.text().lower() == Names[0][0]:
-            self.info_msg(self.tr("Product already exists in database."))
-        else:
-            title = self.ui.title.text().lower()
-            fat = self.ui.fat_value.text()[:-1]
-            satfat = self.ui.sat_fat_value.text()[:-1]
-            carbs = self.ui.carb_value.text()[:-1]
-            sugar = self.ui.sugar_value.text()[:-1]
-            prot = self.ui.protein_value.text()[:-1]
-            salt = self.ui.salt_value.text()[:-1]
-            fiber = self.ui.fiber_value.text()[:-1]
-            self.cursor.execute(f"INSERT INTO products VALUES (\"{title}\", {fat}, {satfat}, {carbs}, {sugar}, {prot}, {salt}, {fiber})")
-            self.database.commit()
-            
-            for combobox in self.ComboBox_tuple:
-                combobox.addItem(self.ui.title.text().lower())
-            self.clear()
-        self.ui.title.setText("")
-        self.ui.title.setPlaceholderText(self.tr("Title"))
 
 def main(VERSION, APPLICATION):
     app = QApplication(sys.argv)